requirements.txt
django==5.1.2
djangorestframework==3.15.2
markdown
drf-yasg
djangorestframework-simplejwt
django-environ
psycopg2-binary